Change parsing of stroke to account for columns
The original idea was to think of the top and bottom rows of the steno
board as the top and bottom rows of a qwerty board, and pressing a
column of keys from the top and bottom row together was to act like the
center row. I wasn't accounting for that in the previous commit, so
pressing two keys was outputting two chars. I changed what I loop
through now, and pull off the bits manually. It seems a little bit slow.
If I wail on the keys, it takes Plover seconds to catch up.

Because pressing two keys together takes a bit of work, I was playing
with the idea of randomizing them, so all 3 options (top, bottom, and
both keys per finger) include more than one row. I'm not sure I love
this now, though.