Change left/right keymaps to map to qwerty better
It's so close to just being a 1:1 match for qwerty, it's a shame I just
don't have a few more keys, like on the Polyglot (where this library
isn't needed/useful). Part of me was thinking I could just add in the
TYGHBN keys by using the HR-FR keys, one at a time, with the star, but
the point of this toy is to allow spamming of keys, i.e. mashing, or
smashing, so for now, I'm staying with the idea of those keys randomly
selecting between the pair of keys I can't get to on my Uni.